refactor(host-application): trim runtime orchestrator facade
Extract run exposure planning into its own module and split workspace runtime startup into smaller helpers. This keeps the AppService facade leaner while preserving the existing runtime startup and run visibility behavior.
